DOMException support + more project restructuring
