+++ index.js
@@ -1,4 +1,9 @@
-import cloneDeep from "./clone-deep.js";
-export { cloneDeepFully, useCustomizers } from "./clone-deep-utils.js";
-export { CLONE } from "./clone-deep-helpers.js";
+import cloneDeep from "./src/clone-deep.js";
+import cloneDeepFully from "./src/clone-deep-fully.js";
+import useCustomizers from "./src/use-customizers.js";
+
 export default cloneDeep;
+export const cloneDeep = cloneDeep;
+export const cloneDeepFully = cloneDeepFully;
+export const useCustomizers = useCustomizers;
+export { CLONE } from "./src/utils.js";
